redis

+++ src/Services/Catalog/Catalog.API/Products/GetProducts/GetProductsEndpoint.cs
@@ -3,7 +3,7 @@ using OpenTelemetry.Trace;
 
 namespace Catalog.API.Products.GetProducts;
 public record GetProductsRequest(int? Page=1,int? PageSize=12,string? Cateloo="") :IQuery<GetProductsResponse>;
-public record GetProductsResponse(IEnumerable<Product> Products,int TotalData =0);
+public record GetProductsResponse(IEnumerable<Product> Products,long TotalData =0);
 public class GetProductsEndpoint : ICarterModule
 {
     public void AddRoutes(IEndpointRouteBuilder app)
@@ -12,15 +12,10 @@ public class GetProductsEndpoint : ICarterModule
         {
            
             var query = request.Adapt<GetProductsQuery>();
-
             var response = await sender.Send(query);
-
             var result = response.Adapt<GetProductsResponse>();
 
             return result;
-     
-
-               
         });
     }
 }
